refactor: artwork API 타입 분리 (artwork.types.ts)
