+++ apps/client/lib/api/artwork.types.ts
@@ -1,0 +1,120 @@
+// ─── 메인 / 전시물 목록 ──────────────────────────────────────────────────────
+
+export interface ArtworkItem {
+	id: string;
+	title: string;
+	category?: string;
+	imageUrl: string;
+	exhibitionTitle: string;
+	exhibitionId: string;
+	slug: string;
+	deptName: string;
+	artistName: string;
+}
+
+export interface ExhibitionMap {
+	id: string;
+	imageUrl: string;
+	description: string | null;
+}
+
+export interface ArtworkArtist {
+	id: string;
+	name: string;
+}
+
+export interface ArtworkListItem {
+	id: string;
+	title: string;
+	imageUrl: string;
+	exhibitionId: string;
+	slug: string;
+	exhibitionTitle: string;
+	artistName: string;
+}
+
+export interface CategoryGroup {
+	categoryName: string;
+	artworks: ArtworkListItem[];
+}
+
+export interface ZoneGroup {
+	zoneName: string;
+	zoneOrderId: number;
+	categories: CategoryGroup[];
+}
+
+export interface ArtworkListResponse {
+	exhibitionId: string;
+	maps: ExhibitionMap[];
+	zones: ZoneGroup[];
+}
+
+export interface GetArtworksParams {
+	zone?: string;
+	category?: string;
+	search?: string;
+}
+
+// ─── 전시물 상세 ──────────────────────────────────────────────────────────────
+
+// 작가 SNS
+export interface ArtistSns {
+	platformName: string;
+	url: string;
+}
+
+// 참여 작가 상세
+export interface ArtworkParticipant {
+	artistId: string;
+	profileId: string;
+	nameKo: string;
+	nameEn: string;
+	profileImg: string | null;
+	role?: string;
+	bio?: string;
+	email?: string | null;
+	sns?: ArtistSns[];
+}
+
+// 상세 이미지
+export interface ArtworkDetailImage {
+	imageUrl: string;
+	description: string;
+}
+
+// BTS
+export interface ArtworkBts {
+	id: string;
+	title: string;
+	mainImg: string;
+	author: string;
+}
+
+// 관련 작품 (sameCategoryArtworks, alphabeticalArtworks 공통)
+export interface RelatedArtwork {
+	id: string;
+	title: string;
+	category: string;
+	artistName: string;
+	mainImage: string;
+	type?: "prev" | "next";
+}
+
+// 작품 상세 전체 응답
+export interface ArtworkDetail {
+	title: string;
+	category: string;
+	material?: string;
+	size?: string | null;
+	description: string;
+	purchaseUrl?: string | null;
+	mainImage: string;
+	locationMap: string;
+	detailImages: ArtworkDetailImage[];
+	youtubeUrl?: string | null;
+	participants: ArtworkParticipant[];
+	relatedBts?: ArtworkBts[];
+	sameCategoryArtworks: RelatedArtwork[];
+	alphabeticalArtworks: RelatedArtwork[];
+}
